[chore] DBI 테스트 KDefineTests로 이동
