+++ Project17-C-Map/InteractiveClusteringMapTests/KMeans/KDefineTests.swift
@@ -16,14 +16,6 @@ class KDefineTests: XCTestCase {
     private let cluster1 = Cluster(coordinates: [Coordinate(x: 1, y: 1), Coordinate(x: 2, y: 2)], boundingBox: .korea)
     private var cluster2 = Cluster(coordinates: [Coordinate(x: -2, y: 1), Coordinate(x: -3, y: -2), Coordinate(x: 0, y: -2)], boundingBox: .korea)
     
-    override func setUpWithError() throws {
-        
-    }
-    
-    override func tearDownWithError() throws {
-        
-    }
-    
     func test_rule_of_thumb_with_5000() throws {
         let expected = 50
         let numberOfData = 5000
@@ -40,6 +32,22 @@ class KDefineTests: XCTestCase {
         let clusters = Clusters(items: [cluster1, cluster2])
         let expected = 0.441
         XCTAssertEqual(expected, clusters.silhouette(), accuracy: accuracy)
+    }
+    
+    private let dbi = DBI()
+    
+    func tests_DBI() throws {
+        let clusters1 = [Cluster(coordinates: [Coordinate(x: 1, y: 1), Coordinate(x: 5, y: 1)], boundingBox: .korea),
+                         Cluster(coordinates: [Coordinate(x: 1, y: 3), Coordinate(x: 5, y: 3)], boundingBox: .korea)]
+        let expected = 2.0
+        XCTAssertEqual(expected, dbi.calculateDBI(clusters: clusters1))
+    }
+    
+    func tests_DBI_2() throws {
+        let clusters2 = [Cluster(coordinates: [Coordinate(x: 1, y: 1), Coordinate(x: 1, y: 3)], boundingBox: .korea),
+                         Cluster(coordinates: [Coordinate(x: 5, y: 1), Coordinate(x: 5, y: 3)], boundingBox: .korea)]
+        let expected = 0.5
+        XCTAssertEqual(expected, dbi.calculateDBI(clusters: clusters2))
     }
     
 }
